Fix Batch1 seed script: copy content from misnamed file and remove duplicate
The file "Seedcr tmh601 batch1 sections1to4.js" (175,342 bytes) had the actual
content, while "seedCR-TMH601-Batch1-Sections1to4.js" was empty (0 bytes).
Copied content to correct filename and removed the misnamed duplicate.
